requirements.txt
attrs==21.2.0
click==8.0.3
jsonschema==4.3.2
paho-mqtt==1.6.1
pip==21.3.1
pyrsistent==0.18.0
PyYAML==6.0
setuptools==58.1.0
